encrypt shop Id

+++ Repository/Shopkeeper/IShopkeeperRepo.cs
@@ -8,6 +8,6 @@ namespace Repository.Shopkeeper
         public Task<List<ShopDetails>> GetShopDetails();
 
         public Task<ShopDetailsById> GetShopDetailsById(long id);
-        public Task<List<ImageModel>> GetShopImageById(int id);
+        public Task<List<ImageModel>> GetShopImageById(long id);
     }
 }
